source init after build_landscape_yaml.sh
When manually calling the build_landscape_yaml.sh script to recreate the landscape.yaml
file after changing something in landscape_config.yaml, there could be inconsistencies
between the yaml file and the values in environment variables, causing the setup to not
work anymore. This is now prevented by always sourcing init.sh when a new landscape.yaml
is created.
Since build_landscape_yaml.sh is called by init.sh if the landscape.yaml file doesn't exist,
this would cause init.sh to be sourced two times in these cases. To prevent that, the
--no-source option for build_landscape_yaml.sh deactivates sourcing init.